RhombusAI/requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
attrs==23.1.0
autopep8==2.0.4
azure-core==1.27.0
azure-storage-blob==12.16.0
beautifulsoup4==4.12.3
boto3==1.34.54
botocore==1.34.54
bs4==0.0.2
cachetools==5.3.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
cmdstanpy==1.2.1
contourpy==1.0.7
coverage==7.4.3
cropyble==1.2.0
cryptography==41.0.1
cycler==0.11.0
Django==4.1.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
dnspython==2.4.2
et-xmlfile==1.1.0
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
google-auth==2.17.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
holidays==0.44
idna==3.4
imageio==2.27.0
importlib_resources==6.1.3
iniconfig==2.0.0
isodate==0.6.1
jax==0.4.7
jmespath==1.0.1
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mediapipe==0.10.7
ml-dtypes==0.0.4
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.8.1.78
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.8.14
packaging==23.0
pandas==1.5.3
patsy==0.5.6
pdf2image==1.16.3
pdftotext==2.2.2
Pillow==9.4.0
plotly==5.13.1
pluggy==1.4.0
prophet==1.1.5
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.11.1
pycparser==2.21
pymongo==4.6.1
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pyparsing==3.0.9
pytesseract==0.3.10
pytest==8.0.2
pytest-cov==4.1.0
python-dateutil==2.8.2
pytz==2023.2
reindent==3.5.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.0
scipy==1.10.1
six==1.16.0
sounddevice==0.4.6
soupsieve==2.5
sqlparse==0.4.3
stanio==0.3.0
statsmodels==0.14.1
tenacity==8.2.2
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
tqdm==4.66.2
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.14.1
